refactor: 게스트 모드 헤더 안내 UI 컴포넌트 분리 #48
